Moved FileUtils to common package
